chore(broker): add metric exporter
The execution metric exporter will collect informations based on the seen records
and publish latency metric like:

 * process execution time, shows the time between creating and completing/terminating a workflow instance
 * job activation time, shows the time between creating an job and activating it
 * job lifetime, show the time between creating an job and completing it

As it makes more sense for short running process and not on all use cases it is disabled by default.
We can enabled it for example with a env variable, which we will do with
out benchmark setup. For performance analysis the metrics are quite
interesting. We moved these metrics into an exporter, such that we not
have this in our current state (db) since it is not always needed and
can easily be extended.